Move template overrides to new class
